+++ Sofa/Component/Collision/Detection/Algorithm/src/sofa/component/collision/detection/algorithm/RayTraceDetection.h
@@ -33,13 +33,13 @@ namespace sofa::component::collision::detection::algorithm
  *
  *   For each point in one object, we trace a ray following the opposite of the point's normal
  *   up to find a triangle in the other object. Both triangles are tested to evaluate if they are in
- *   colliding state. It must be used with a TriangleOctreeModel,as an octree is used to traverse the object.
+ *   colliding state. It must be used with a TriangleOctreeCollisionModel,as an octree is used to traverse the object.
  */
 class SOFA_COMPONENT_COLLISION_DETECTION_ALGORITHM_API RayTraceDetection final :
-    public sofa::core::objectmodel::BaseObject
+    public sofa::core::objectmodel::BaseComponent
 {
 public:
-    SOFA_CLASS(RayTraceDetection, sofa::core::objectmodel::BaseObject);
+    SOFA_CLASS(RayTraceDetection, sofa::core::objectmodel::BaseComponent);
 
     void init() override;