+++ Sofa/Component/LinearSystem/src/sofa/component/linearsystem/BaseMatrixProjectionMethod.h
@@ -22,7 +22,7 @@
 #pragma once
 
 #include <sofa/component/linearsystem/config.h>
-#include <sofa/component/linearsystem/MappingGraph.h>
+#include <sofa/simulation/MappingGraph.h>
 #include <sofa/core/behavior/StateAccessor.h>
 #include <sofa/linearalgebra/CompressedRowSparseMatrix.h>
 
@@ -62,7 +62,7 @@ public:
      * \param globalMatrix The product is added into this matrix
      */
     virtual void projectMatrixToGlobalMatrix(const core::MechanicalParams* mparams,
-                                             const MappingGraph& mappingGraph,
+                                             const simulation::MappingGraph& mappingGraph,
                                              TMatrix* matrixToProject,
                                              linearalgebra::BaseMatrix* globalMatrix) = 0;