+++ Sofa/GL/Component/Shader/tests/Light_test.cpp
@@ -101,7 +101,7 @@ void TestLight::checkLightMissingLightManager(const std::string& lighttype)
     ASSERT_NE(root.get(), nullptr) ;
     root->init(sofa::core::execparams::defaultInstance()) ;
 
-    sofa::core::objectmodel::BaseObject* lm = root->getTreeNode("Level 1")->getObject("light1") ;
+    sofa::core::objectmodel::BaseComponent* lm = root->getTreeNode("Level 1")->getObject("light1") ;
     ASSERT_NE(lm, nullptr) ;
 
     sofa::simulation::node::unload(root);
@@ -126,10 +126,10 @@ void TestLight::checkPositionalLightValidAttributes()
     ASSERT_NE(root.get(), nullptr) ;
     root->init(sofa::core::execparams::defaultInstance()) ;
 
-    sofa::core::objectmodel::BaseObject* lm = root->getTreeNode("Level 1")->getObject("lightmanager") ;
+    sofa::core::objectmodel::BaseComponent* lm = root->getTreeNode("Level 1")->getObject("lightmanager") ;
     ASSERT_NE(lm, nullptr) ;
 
-    sofa::core::objectmodel::BaseObject* light = root->getTreeNode("Level 1")->getObject("light1") ;
+    sofa::core::objectmodel::BaseComponent* light = root->getTreeNode("Level 1")->getObject("light1") ;
     ASSERT_NE(light, nullptr) ;
 
     /// List of the supported attributes the user expect to find
@@ -167,10 +167,10 @@ void TestLight::checkDirectionalLightValidAttributes()
     ASSERT_NE(root.get(), nullptr) ;
     root->init(sofa::core::execparams::defaultInstance()) ;
 
-    sofa::core::objectmodel::BaseObject* lm = root->getTreeNode("Level 1")->getObject("lightmanager") ;
+    sofa::core::objectmodel::BaseComponent* lm = root->getTreeNode("Level 1")->getObject("lightmanager") ;
     ASSERT_NE(lm, nullptr) ;
 
-    sofa::core::objectmodel::BaseObject* light = root->getTreeNode("Level 1")->getObject("light1") ;
+    sofa::core::objectmodel::BaseComponent* light = root->getTreeNode("Level 1")->getObject("light1") ;
     ASSERT_NE(light, nullptr) ;
 
     /// List of the supported attributes the user expect to find
@@ -208,10 +208,10 @@ void TestLight::checkSpotLightValidAttributes()
     ASSERT_NE(root.get(), nullptr) ;
     root->init(sofa::core::execparams::defaultInstance()) ;
 
-    sofa::core::objectmodel::BaseObject* lm = root->getTreeNode("Level 1")->getObject("lightmanager") ;
+    sofa::core::objectmodel::BaseComponent* lm = root->getTreeNode("Level 1")->getObject("lightmanager") ;
     ASSERT_NE(lm, nullptr) ;
 
-    sofa::core::objectmodel::BaseObject* light = root->getTreeNode("Level 1")->getObject("light1") ;
+    sofa::core::objectmodel::BaseComponent* light = root->getTreeNode("Level 1")->getObject("light1") ;
     ASSERT_NE(light, nullptr) ;
 
     /// List of the supported attributes the user expect to find