+++ Sofa/framework/Core/src/sofa/core/objectmodel/BaseData.h
@@ -291,11 +291,14 @@ public:
     template<class T>
     static std::string typeName()
     {
-        if (defaulttype::DataTypeInfo<T>::ValidInfo)
+        if constexpr (defaulttype::DataTypeInfo<T>::ValidInfo)
         {
             return defaulttype::DataTypeName<T>::name();
         }
-        return decodeTypeName(typeid(T));
+        else
+        {
+            return decodeTypeName(typeid(T));
+        }
     }
 
 protected: