+++ Sofa/framework/Geometry/src/sofa/geometry/Prism.h
@@ -33,6 +33,19 @@ struct Prism
     static constexpr ElementType Element_type = ElementType::PRISM;
 
     Prism() = delete;
+
+    /* CONVENTION : indices ordering for the nodes of a prism :
+     *
+     *      5
+     *    / |  \
+     *   /  |   \
+     *  3---+----4
+     *  |   |    |
+     *  |   2    |
+     *  | /   \  |
+     *  |/      \|
+     *  0--------1
+     */
 };
 
 using Pentahedron SOFA_ATTRIBUTE_DEPRECATED("v25.12", "v26.06", "Pentahedron is renamed to Prism") = Prism;