+++ Sofa/framework/Simulation/Common/src/sofa/simulation/common/xml/AttributeElement.cpp
@@ -29,7 +29,7 @@ using namespace sofa::defaulttype;
 using helper::Creator;
 
 AttributeElement::AttributeElement(const std::string& name, const std::string& type, BaseElement* eparent)
-    : Element<core::objectmodel::BaseObject>(name, type, eparent)
+    : Element<core::objectmodel::BaseComponent>(name, type, eparent)
 {
 }