+++ Sofa/framework/Simulation/Core/src/sofa/simulation/BaseMechanicalVisitor.cpp
@@ -53,6 +53,22 @@ BaseMechanicalVisitor::BaseMechanicalVisitor(const sofa::core::ExecParams *param
     canAccessSleepingNode = false;
 }
 
+bool BaseMechanicalVisitor::isMechanicalStateMapped(simulation::Node* node) const
+{
+    /**
+     * WARNING!!!!!
+     *
+     * This condition is not valid for all cases, as it only checks for the presence of a mechanical
+     * mapping in the same Node. This condition does not ensure that the state is not an output of a
+     * mapping.
+     *
+     * Examples of situations where this condition is not appropritate:
+     * - No mapping in the Node but the mapping is defined in another Node.
+     * - There is a mapping in the Node, but the state is not an output of this mapping.
+     */
+    return node->mechanicalMapping != nullptr;
+}
+
 Visitor::Result BaseMechanicalVisitor::processNodeTopDown(simulation::Node *node, VisitorContext *ctx)
 {
     for (auto *solver : node->solver)
@@ -76,7 +92,7 @@ Visitor::Result BaseMechanicalVisitor::processNodeTopDown(simulation::Node *node
 
     if (node->mechanicalState != nullptr)
     {
-        if (node->mechanicalMapping != nullptr)
+        if (isMechanicalStateMapped(node))
         {
             res = runVisitorTask(this, ctx, &BaseMechanicalVisitor::fwdMappedMechanicalState, &*node->mechanicalState, fwdVisitorType);
         }
@@ -134,7 +150,7 @@ void BaseMechanicalVisitor::processNodeBottomUp(simulation::Node *node, VisitorC
 
     if (node->mechanicalState != nullptr)
     {
-        if (node->mechanicalMapping != nullptr)
+        if (isMechanicalStateMapped(node))
         {
             runVisitorTask(this, ctx, &BaseMechanicalVisitor::bwdMappedMechanicalState, &*node->mechanicalState, bwdVisitorType);
         }
@@ -225,7 +241,7 @@ void BaseMechanicalVisitor::printWriteVectors(core::behavior::BaseMechanicalStat
 }
 
 
-void BaseMechanicalVisitor::printReadVectors(simulation::Node* node, core::objectmodel::BaseObject* obj)
+void BaseMechanicalVisitor::printReadVectors(simulation::Node* node, core::objectmodel::BaseComponent* obj)
 {
     using sofa::core::behavior::BaseInteractionForceField;
     using sofa::core::behavior::BaseInteractionProjectiveConstraintSet;
@@ -272,7 +288,7 @@ void BaseMechanicalVisitor::printReadVectors(simulation::Node* node, core::objec
 }
 
 
-void BaseMechanicalVisitor::printWriteVectors(simulation::Node* node, core::objectmodel::BaseObject* obj)
+void BaseMechanicalVisitor::printWriteVectors(simulation::Node* node, core::objectmodel::BaseComponent* obj)
 {
     using sofa::core::behavior::BaseInteractionForceField;
     using sofa::core::behavior::BaseInteractionProjectiveConstraintSet;
@@ -323,7 +339,7 @@ void BaseMechanicalVisitor::printWriteVectors(simulation::Node* node, core::obje
 }
 
 
-Visitor::ctime_t BaseMechanicalVisitor::begin(simulation::Node* node, core::objectmodel::BaseObject* obj, const std::string &info)
+Visitor::ctime_t BaseMechanicalVisitor::begin(simulation::Node* node, core::objectmodel::BaseComponent* obj, const std::string &info)
 {
     ctime_t t=Visitor::begin(node, obj, info);
     printReadVectors(node, obj);
@@ -331,7 +347,7 @@ Visitor::ctime_t BaseMechanicalVisitor::begin(simulation::Node* node, core::obje
 }
 
 
-void BaseMechanicalVisitor::end(simulation::Node* node, core::objectmodel::BaseObject* obj, ctime_t t0)
+void BaseMechanicalVisitor::end(simulation::Node* node, core::objectmodel::BaseComponent* obj, ctime_t t0)
 {
     printWriteVectors(node, obj);
     Visitor::end(node, obj, t0);